// map-your-value/front/src/type.tsx
export type Competitors = {
    Competitor: string;
    Descriptive_summary: string;
    Strengths: string[] | string;
    Weaknesses: string[] | string;
    Proximity_score: number;
    Proximity_Explanation: string;
    Crunchbase_Link: string;
    Company_Card: string;
};



export type Features = {
    feature: string;
};